scripts/requirements.txt
pyyaml
gitpython
unflatten